backend/requirements.txt
flask==3.0.*
flask-cors==4.0.*
mysql-connector-python==9.0.*
pytest==8.0.*
